refactor(page-index): SSR 이 자기 자신을 HTTP 로 부르던 계층 제거
SSR([boardId]/[postId]/+page.server.ts)이 svelteKitFetch 로
/api/boards/.../page-index 를 호출하고 있었다. 같은 프로세스라 네트워크는 없지만
Request/Response 생성과 JSON 직렬화·역직렬화가 매 요청 일어난다.
캐시 적중인데도 1~3ms 가 걸린 이유다(Redis GET 하나면 1ms 미만이어야 한다).

이 계산의 호출자는 100% SSR 이다. 2026-08-18 실측:
운영 파드 nginx 사이드카 로그 3,000줄에서 브라우저 API 2,230건
(좋아요 233·댓글 154·likers-batch 109) 중 page-index 요청은 0건이었다.
즉 HTTP 계층은 순수 낭비였다.

변경:
- 계산 로직을 /server/page-index.ts 로 추출
- 라우트는 얇은 래퍼로 축소 (공개 계약·응답 스키마·Cache-Control 동일)
- SSR 은 getPageIndex() 를 직접 호출

동일하게 유지한 것:
- Redis 키 (pi:prior:{board}:{postId}, pi:rows:{board})
- 깊이별 TTL (60s / 600s / 3600s), page_rows TTL 600s
- 응답 스키마 {page, page_rows, prior}
- 실패 시 page:1 fallback (getPageIndex 는 던지지 않는다)

공개 라우트는 지우지 않았다. recent-posts.svelte 에 폴백 호출이 남아 있다
(SSR 이 1페이지를 준 경우에만 발화).

부수 효과: 요청당 객체 할당 감소. 2026-08-07 web heap OOM 이력이 있는 서비스다.
